+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Added
   * CLI command `ocrd-tool`
   * CLI command `workspace find`
   * CLI command `workspace create`
+  * CLI `workspace` flag --no-cache
   * bash library to build wrappers
 Changed
 Fixed